+++ src/main/resources/locale/en.properties
@@ -132,3 +132,7 @@ missing-ban-perm = I can't do that because I'm missing the Ban permission
 missing-kick-perm = I can't do that because I'm missing the Kick permission
 missing-messages-manage-perm = I can't do that because I'm missing the Manage Chat permission
 missing-voice-perm = I can't do that because I'm missing the Voice Connect permission
+nothing-playing = Couldn't skip the track because nothing is playing
+voteskip-help = Usage: .voteskip - Start a vote to skip the current track
+song-help = Usage: .horosong - Get information about the current song
+translate-help = Usage: .horotranslate [from] [to] [text] - Translate something from one language to another
