+++ src/main/resources/locale/nl.properties
@@ -132,3 +132,7 @@ missing-ban-perm = Ik kan dat helaas niet doen omdat ik de Ban permissie niet he
 missing-kick-perm = Ik kan dat helaas niet doen omdat ik de Kick permissie niet heb
 missing-messages-manage-perm = Ik kan dat helaas niet doen omdat ik de Manage Messages permissie niet heb
 missing-voice-perm = Ik kan dat helaas niet doen omdat ik de Voice Connect permissie niet heb
+nothing-playing = Kon het nummer niet overslaan omdat er geen nummer aan het spelen is
+voteskip-help = Gebruik: .horovoteskip - Stem om het nummer over te slaan
+song-help = Gebruik: .horosong - Krijg informatie over het nummer dat op dit moment aan het afspelen is
+translate-help = Gebruik: .horotranslate [van] [naar] [tekst] - Vertaal iets van een bepaalde taal naar een andere taal
